Rename renderer_wrapper to renderer_context
