+++ ultramodern/CMakeLists.txt
@@ -16,7 +16,7 @@ add_library(ultramodern STATIC
     "${CMAKE_CURRENT_SOURCE_DIR}/src/input.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/mesgqueue.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/misc_ultra.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/renderer_wrapper.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/renderer_context.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/rsp.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/scheduling.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/task_win32.cpp"
